bpf: tests: condense nodeport ClusterIP test
Run the IPv4 and IPv6 test from a single file.
