+++ bpf/tests/tc_lb_clusterip.c
@@ -5,19 +5,25 @@
 #include "common.h"
 #include "pktgen.h"
 
-#define ENABLE_IPV4
-#define ENABLE_NODEPORT
+#define ENABLE_IPV4		1
+#define ENABLE_IPV6		1
+#define ENABLE_NODEPORT		1
 
 #define CLIENT_IP		v4_ext_one
+#define CLIENT_IPV6		v6_ext_node_one
 #define CLIENT_PORT		tcp_src_one
 
 #define FRONTEND_IP		v4_svc_one
+#define FRONTEND_IPV6		v6_svc_one
 #define FRONTEND_PORT		tcp_svc_one
 
 #define BACKEND_IP		v4_pod_one
+#define BACKEND_IPV6		v6_pod_one
 #define BACKEND_PORT		tcp_dst_one
 
 #define NAT_REV_INDEX		1
+#define NAT_REV_INDEX_2		2
+
 #define BACKEND_COUNT		1
 #define BACKEND_IFINDEX		11
 #define BACKEND_ID		124
@@ -38,6 +44,16 @@ const __u8 lb4_clusterip[] = {
 /* packet defined in ./scapy/lb_pkt_defs.py */
 const __u8 lb4_clusterip_post_dnat[] = {
 	SCAPY_BUF_BYTES(lb4_clusterip_post_dnat)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_clusterip[] = {
+	SCAPY_BUF_BYTES(lb6_clusterip)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_clusterip_post_dnat[] = {
+	SCAPY_BUF_BYTES(lb6_clusterip_post_dnat)
 };
 
 /* Test that a request from an external client to a ClusterIP service w/o the
@@ -169,6 +185,139 @@ int lb4_routable_clusterip_check(__maybe_unused const struct __ctx_buff *ctx)
 
 	ASSERT_CTX_BUF_OFF("lb4_routable_clusterip", "Ether", ctx, sizeof(__u32),
 			   lb4_clusterip_post_dnat, sizeof(lb4_clusterip_post_dnat));
+
+	test_finish();
+}
+
+/* Test that a request from an external client to a ClusterIP service w/o the
+ * `bpf-lb-external-clusterip` flag set is denied. The reason is due to the SVC
+ * being created w/o the SVC_FLAG_ROUTABLE flag being set in the bpf map,
+ * leading to the datapath dropping the packet with the reason code
+ * DROP_IS_CLUSTER_IP, and the respective BPF metric updated accordingly.
+ */
+PKTGEN("tc", "tc_lb6_nonroutable_clusterip")
+int lb6_nonroutable_clusterip_pktgen(struct __ctx_buff *ctx)
+{
+	struct pktgen builder;
+
+	pktgen__init(&builder, ctx);
+
+	scapy_push_data(&builder, lb6_clusterip, sizeof(lb6_clusterip));
+
+	pktgen__finish(&builder);
+
+	return 0;
+}
+
+SETUP("tc", "tc_lb6_nonroutable_clusterip")
+int lb6_nonroutable_clusterip_setup(struct __ctx_buff *ctx)
+{
+	endpoint_v6_add_entry((union v6addr *)BACKEND_IPV6, BACKEND_IFINDEX, 0, 0, 0,
+			      (__u8 *)mac_one, (__u8 *)mac_two);
+	lb_v6_add_service_with_flags((union v6addr *)FRONTEND_IPV6, FRONTEND_PORT,
+				     IPPROTO_TCP, BACKEND_COUNT, NAT_REV_INDEX_2,
+				     0, 0);
+	lb_v6_add_backend((union v6addr *)FRONTEND_IPV6, FRONTEND_PORT, BACKEND_COUNT, BACKEND_ID,
+			  (union v6addr *)BACKEND_IPV6, BACKEND_PORT, IPPROTO_TCP, 0);
+
+	return netdev_receive_packet(ctx);
+}
+
+CHECK("tc", "tc_lb6_nonroutable_clusterip")
+int lb6_nonroutable_clusterip_check(__maybe_unused const struct __ctx_buff *ctx)
+{
+	void *data, *data_end;
+	__u32 *status_code, drop_reason;
+	__u64 count = 2;
+	struct metrics_key key = {
+		.reason = -DROP_IS_CLUSTER_IP,
+		.dir = METRIC_INGRESS,
+	};
+
+	test_init();
+
+	data = ctx_data(ctx);
+	data_end = ctx_data_end(ctx);
+	status_code = data;
+	/* Retrieve drop reason from metadata (bpf/lib/drop.h) */
+	drop_reason = ctx_load_meta(ctx, 2);
+
+	assert(data + sizeof(__u32) <= data_end);
+
+	assert(*status_code == CTX_ACT_DROP);
+
+	assert(drop_reason == -DROP_IS_CLUSTER_IP);
+
+	assert_metrics_count(key, count);
+
+	ASSERT_CTX_BUF_OFF("lb6_nonroutable_clusterip", "Ether", ctx, sizeof(__u32),
+			   lb6_clusterip, sizeof(lb6_clusterip));
+
+	test_finish();
+}
+
+/* Test that a request from an external client to a routable ClusterIP service
+ * is allowed. The reason is due to the SVC being created with the SVC_FLAG_ROUTABLE
+ * flag set, leading to the datapath correctly accepting and DNAT the packet.
+ */
+PKTGEN("tc", "tc_lb6_routable_clusterip")
+int lb6_routable_clusterip_pktgen(struct __ctx_buff *ctx)
+{
+	struct pktgen builder;
+
+	pktgen__init(&builder, ctx);
+
+	scapy_push_data(&builder, lb6_clusterip, sizeof(lb6_clusterip));
+
+	pktgen__finish(&builder);
+
+	return 0;
+}
+
+SETUP("tc", "tc_lb6_routable_clusterip")
+int lb6_routable_clusterip_setup(struct __ctx_buff *ctx)
+{
+	/* Endpoint and backend added in previous setup, simply change SVC flags. */
+	lb_v6_add_service_with_flags((union v6addr *)FRONTEND_IPV6, FRONTEND_PORT,
+				     IPPROTO_TCP, BACKEND_COUNT, NAT_REV_INDEX_2,
+				     SVC_FLAG_ROUTABLE, 0);
+
+	return netdev_receive_packet(ctx);
+}
+
+CHECK("tc", "tc_lb6_routable_clusterip")
+int lb6_routable_clusterip_check(__maybe_unused const struct __ctx_buff *ctx)
+{
+	void *data, *data_end;
+	__u32 *status_code;
+	struct ct_entry *ct_entry;
+	struct ipv6_ct_tuple tuple = {
+		.saddr   = *(union v6addr *)CLIENT_IPV6,
+		.daddr   = *(union v6addr *)FRONTEND_IPV6,
+		.sport   = FRONTEND_PORT,
+		.dport   = CLIENT_PORT,
+		.nexthdr = IPPROTO_TCP,
+		.flags   = TUPLE_F_SERVICE,
+	};
+
+	test_init();
+
+	data = ctx_data(ctx);
+	data_end = ctx_data_end(ctx);
+	status_code = data;
+
+	assert(data + sizeof(__u32) <= data_end);
+
+	assert(*status_code == CTX_ACT_OK);
+
+	/* Ensure CT entry is updated accordingly (SVC). */
+	ct_entry = map_lookup_elem(get_ct_map6(&tuple), &tuple);
+	assert(ct_entry);
+	assert(ct_entry->packets == 1);
+	assert(ct_entry->bytes == ctx_full_len(ctx) - sizeof(__u32));
+
+	ASSERT_CTX_BUF_OFF("lb6_routable_clusterip", "Ether", ctx, sizeof(__u32),
+			   lb6_clusterip_post_dnat, sizeof(lb6_clusterip_post_dnat));
 
 	test_finish();
 }
